api/requirements.txt
Django==2.2.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-hosts==3.0
djangorestframework==3.9.4
ipython==7.3.0
psycopg2==2.7.3.2
